requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.9.0
azure-cosmos==4.2.0
azure-cli-core==2.34.1
pymongo==4.1.1
Flask==1.1.2
requests==2.23.0